requirements.txt
a2wsgi==1.10.4
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.6
exceptiongroup==1.2.0
Flask==3.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1
